fix(plugins): init - lock impatient commit
There is a small issue upstream, which errors out plugin init.  This
isnt a big deal, but for now (very temp, as the new nightly NeoVim build
should resolve) we are locking to the previous working commit.

Will revert this commmit as soon as the fresh appimages & mac binaries
are created by NeoVim ci/cd processes.